Display errors for class expressions (#70)
The current behavior of the `no-class` rule is to display errors for
class declarations but not for class expressions. This alters the
rule to display errors for both cases, fixing issue #65. I can't
think of a instance of a user wanting one but not the other, so I
didn't introduce a option for specifying so.